Reorganize `trace` submodule
## Refactoring: reorganize trace module package

- **Package restructuring:** extract from `trace` package several smaller sub-packages:
    - `trace.printing` -- trace printing utilities;
    - `trace.serialization` -- (binary) serialization logic;
    - `trace.storage` helper utility storage classes for binary serialization.
    
- **Extract utility collections:** create `util.collections` sub-package and extract utility collections classes there:
    - `Bitmap`;
    - `SortedList`; 
    - `ChunkedList`; 
    - `Collections.kt` -- util functions.
     
- Update all imports across the codebase accordingly. 

---------
