+++ common/src/main/org/jetbrains/lincheck/util/collections/Collections.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.util
+package org.jetbrains.lincheck.util.collections
 
 import java.util.Collections
 import java.util.IdentityHashMap
