+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/eventstructure/EventAggregation.kt
@@ -21,12 +21,8 @@
 package org.jetbrains.kotlinx.lincheck.strategy.managed.eventstructure
 
 import org.jetbrains.kotlinx.lincheck.util.*
-import org.jetbrains.lincheck.util.Covering
-import org.jetbrains.lincheck.util.Relation
-import org.jetbrains.lincheck.util.ensure
-import org.jetbrains.lincheck.util.ensureNotNull
-import org.jetbrains.lincheck.util.ensureNull
-import org.jetbrains.lincheck.util.squash
+import org.jetbrains.lincheck.util.*
+import org.jetbrains.lincheck.util.collections.*
 
 interface EventAggregator {
     fun aggregate(events: List<AtomicThreadEvent>): List<List<AtomicThreadEvent>>
