+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/eventstructure/EventIndexing.kt
@@ -21,15 +21,8 @@
 package org.jetbrains.kotlinx.lincheck.strategy.managed.eventstructure
 
 import org.jetbrains.kotlinx.lincheck.strategy.managed.*
-import org.jetbrains.kotlinx.lincheck.util.*
-import org.jetbrains.lincheck.util.Enumerator
-import org.jetbrains.lincheck.util.SortedList
-import org.jetbrains.lincheck.util.SortedMutableList
-import org.jetbrains.lincheck.util.sortedListOf
-import org.jetbrains.lincheck.util.sortedMutableListOf
-import org.jetbrains.lincheck.util.toEnumerator
-import org.jetbrains.lincheck.util.unordered
-import org.jetbrains.lincheck.util.updateInplace
+import org.jetbrains.lincheck.util.*
+import org.jetbrains.lincheck.util.collections.*
 
 
 typealias EventIndexClassifier<E, C, K> = (E) -> Pair<C, K>?
