+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/eventstructure/ExtendedExecution.kt
@@ -24,7 +24,7 @@ import org.jetbrains.kotlinx.lincheck.strategy.managed.eventstructure.consistenc
 import org.jetbrains.kotlinx.lincheck.util.*
 import org.jetbrains.lincheck.util.ComputableNode
 import org.jetbrains.lincheck.util.Relation
-import org.jetbrains.lincheck.util.SortedList
+import org.jetbrains.lincheck.util.collections.SortedList
 import org.jetbrains.lincheck.util.computable
 import org.jetbrains.lincheck.util.dependsOn
 import org.jetbrains.lincheck.util.union
