+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/eventstructure/consistency/ReadModifyWriteAtomicityChecker.kt
@@ -22,11 +22,8 @@ package org.jetbrains.kotlinx.lincheck.strategy.managed.eventstructure.consisten
 
 import org.jetbrains.kotlinx.lincheck.strategy.managed.*
 import org.jetbrains.kotlinx.lincheck.strategy.managed.eventstructure.*
-import org.jetbrains.lincheck.util.Computable
-import org.jetbrains.lincheck.util.Relation
-import org.jetbrains.lincheck.util.ensure
-import org.jetbrains.lincheck.util.refine
-import org.jetbrains.lincheck.util.updateInplace
+import org.jetbrains.lincheck.util.*
+import org.jetbrains.lincheck.util.collections.*
 
 // TODO: restore atomicity violation information
 class ReadModifyWriteAtomicityViolation(/*val write1: Event, val write2: Event*/) : Inconsistency() {
