+++ trace/src/main/org/jetbrains/lincheck/trace/TraceRecorderTracePoints.kt
@@ -11,15 +11,19 @@
 package org.jetbrains.lincheck.trace
 
 import org.jetbrains.lincheck.descriptors.*
-import org.jetbrains.lincheck.trace.DefaultTRArrayTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRCatchTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRFieldTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRLineBreakpointSnapshotTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRLocalVariableTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRLoopIterationTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRLoopTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRMethodCallTracePointPrinter.append
-import org.jetbrains.lincheck.trace.DefaultTRThrowTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.*
+import org.jetbrains.lincheck.trace.serialization.*
+import org.jetbrains.lincheck.trace.storage.AddressIndex
+import org.jetbrains.lincheck.util.collections.ChunkedList
+import org.jetbrains.lincheck.trace.printing.DefaultTRArrayTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRCatchTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRFieldTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRLineBreakpointSnapshotTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRLocalVariableTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRLoopIterationTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRLoopTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRMethodCallTracePointPrinter.append
+import org.jetbrains.lincheck.trace.printing.DefaultTRThrowTracePointPrinter.append
 import java.io.DataInput
 import java.io.DataOutput
 import java.math.BigDecimal
