+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/Deserialization.kt
@@ -8,9 +8,10 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.trace
+package org.jetbrains.lincheck.trace.serialization
 
-import org.jetbrains.lincheck.descriptors.*
+import org.jetbrains.lincheck.trace.*
+import org.jetbrains.lincheck.trace.storage.RangeIndex
 import org.jetbrains.lincheck.util.Logger
 import java.nio.file.StandardOpenOption
 import java.nio.file.Files
