+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/NetworkSerialization.kt
@@ -8,8 +8,9 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.trace
+package org.jetbrains.lincheck.trace.serialization
 
+import org.jetbrains.lincheck.trace.*
 import org.jetbrains.lincheck.trace.network.TracingCallbacks
 import org.jetbrains.lincheck.trace.network.TracingServer
 import org.jetbrains.lincheck.util.Logger
