+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/SerializationInmemory.kt
@@ -8,9 +8,11 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.trace
+package org.jetbrains.lincheck.trace.serialization
 
+import org.jetbrains.lincheck.trace.*
 import org.jetbrains.lincheck.util.Logger
+import org.jetbrains.lincheck.util.collections.SimpleBitmap
 import java.io.DataOutput
 import java.io.DataOutputStream
 import java.io.OutputStream
