+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/SerializationPack.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.trace
+package org.jetbrains.lincheck.trace.serialization
 
 import org.jetbrains.lincheck.util.Logger
 import java.io.DataInputStream
