+++ trace/src/main/org/jetbrains/lincheck/trace/storage/MemMapTemporaryStorage.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck.trace
+package org.jetbrains.lincheck.trace.storage
 
 import sun.nio.ch.DirectBuffer
 import java.io.Closeable
